Start splitting init tests.
