+++ test/server/spec/bin/builder-init.spec.js
@@ -1,0 +1,1 @@
+"use strict";
